Visual C++ 6 fixes.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3824 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_io.c
@@ -531,7 +531,7 @@ mx_set_file_offset( FILE *mx_file, int64_t offset, int origin )
 		return (-1);
 	}
 
-	offset = fseek( mx_file, offset, origin );
+	offset = fseek( mx_file, (long) offset, origin );
 
 	return offset;
 }
